pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Melkoh <milko.gutierrez@innovatica.com.py>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
dj-rest-auth = "^6.0.0"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
django-cities-light = "^3.10.1"
faker = "^25.3.0"
django-jazzmin = "^3.0.0"
psycopg2 = "^2.9.9"
geoip2 = "^4.8.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
